refactor: move contributor orderBy validation to schema gf-504
