+++ apps/backend/src/modules/contributors/libs/validation-schemas/validation-schemas.ts
@@ -1,4 +1,5 @@
 export {
+	contributorGetAllValidationSchema,
 	contributorMergeValidationSchema,
 	contributorPatchValidationSchema,
 	contributorSplitValidationSchema,
