+++ packages/shared/src/modules/contributors/libs/enums/contributor-validation-message.enum.ts
@@ -2,6 +2,7 @@ import { ContributorValidationRule } from "./contributor-validation-rule.enum.js
 
 const ContributorValidationMessage = {
 	GIT_EMAIL_REQUIRED: "Git email is required.",
+	INVALID_ORDER_BY_QUERY_PARAMETER: "Invalid orderBy query parameter.",
 	NAME_REQUIRED: "Contributor name is required.",
 	NAME_TOO_LONG: `Name is too long (> ${String(ContributorValidationRule.NAME_MAXIMUM_LENGTH)} symbols).`,
 	NAME_TOO_SHORT: `Name must be at least ${String(ContributorValidationRule.NAME_MINIMUM_LENGTH)} characters long.`,
