+++ packages/shared/src/modules/contributors/libs/enums/enums.ts
@@ -1,4 +1,4 @@
-export { ContributorOrderByKey } from "./contributor-order-by.enum.js";
+export { ContributorOrderByKey } from "./contributor-order-by-key.enum.js";
 export { ContributorValidationMessage } from "./contributor-validation-message.enum.js";
 export { ContributorValidationRule } from "./contributor-validation-rule.enum.js";
 export { ContributorsApiPath } from "./contributors-api-path.enum.js";
