+++ packages/shared/src/modules/contributors/libs/types/contributor-get-all-query-parameters.type.ts
@@ -2,12 +2,12 @@ import {
 	type PaginationQueryParameters,
 	type ValueOf,
 } from "../../../../libs/types/types.js";
-import { type ContributorOrderByKey } from "../enums/contributor-order-by.enum.js";
+import { type ContributorOrderByKey } from "../enums/contributor-order-by-key.enum.js";
 
 type ContributorGetAllQueryParameters = {
 	contributorName?: string | undefined;
 	hasHidden?: boolean | undefined;
-	orderBy?: ValueOf<typeof ContributorOrderByKey>;
+	orderBy?: undefined | ValueOf<typeof ContributorOrderByKey>;
 	projectId?: number | undefined;
 } & Partial<PaginationQueryParameters>;
