+++ packages/shared/src/modules/contributors/libs/validation-schemas/validation-schemas.ts
@@ -1,3 +1,4 @@
+export { contributorGetAll } from "./contributor-get-all.validation-schema.js";
 export { contributorMerge } from "./contributor-merge.validation-schema.js";
 export { contributorPatch } from "./contributor-patch.validation-schema.js";
 export { contributorSplit } from "./contributor-split.validation-schema.js";
